py/requirements.txt
pyautogui==0.9.54
pyobjc-core>=9.2
pyobjc-framework-Quartz>=9.2
Pillow>=10.0
